# More performance regressions