--- a/kore/src/main/scala/org/kframework/definition/outer.scala
+++ b/kore/src/main/scala/org/kframework/definition/outer.scala
@@ -368,26 +368,6 @@ case class Module(val name: String, val imports: Set[Import], localSentences: Se
     case _ =>
   }
 
-  def checkAtts(other: Set[Sentence]): Boolean = {
-    if (sentences != other)
-      return false
-
-    val productionsSorted = sentences.collect({ case p: Production => p }).toSeq.sorted
-    val otherProductionSorted = other.collect({ case p: Production => p }).toSeq.sorted
-
-    if (productionsSorted != otherProductionSorted)
-      return false
-
-    for (i <- productionsSorted.indices) {
-      val p1 = productionsSorted(i)
-      val p2 = otherProductionSorted(i)
-      if (!p1.att.equals(p2.att))
-        return false
-    }
-
-    true
-  }
-
   def checkUserLists(): Unit = localSentences foreach {
     case p@Production(_, _, srt, _, atts) =>
       if (atts.contains(Att.USER_LIST)) {