@radumereuta 2023-08-04
> LGTM as these are just clean reverts. Make sure to open an issue for investigation later as well!

Fair point. Added a comment in the issue that keeps track of these changes: https://github.com/runtimeverification/k/issues/3444#issuecomment-1665461900